requirements.txt
configparser
urllib3
PyYAML
requests
termcolor
